+++ database/ent/rolepermission.go
@@ -23,15 +23,15 @@ type RolePermission struct {
 	// 创建时间
 	CreateTime time.Time `json:"create_time,omitempty"`
 	// 创建人ID
-	CreateBy int64 `json:"create_by,omitempty"`
+	CreateBy uint64 `json:"create_by,omitempty"`
 	// 更新时间
 	UpdateTime time.Time `json:"update_time,omitempty"`
 	// 更新人ID
-	UpdateBy int64 `json:"update_by,omitempty"`
+	UpdateBy uint64 `json:"update_by,omitempty"`
 	// 删除时间
 	DeleteTime time.Time `json:"delete_time,omitempty"`
 	// 删除人ID
-	DeleteBy int64 `json:"delete_by,omitempty"`
+	DeleteBy uint64 `json:"delete_by,omitempty"`
 	// RoleID holds the value of the "role_id" field.
 	RoleID uint64 `json:"role_id,omitempty"`
 	// PermissionID holds the value of the "permission_id" field.
@@ -115,7 +115,7 @@ func (_m *RolePermission) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field create_by", values[i])
 			} else if value.Valid {
-				_m.CreateBy = value.Int64
+				_m.CreateBy = uint64(value.Int64)
 			}
 		case rolepermission.FieldUpdateTime:
 			if value, ok := values[i].(*sql.NullTime); !ok {
@@ -127,7 +127,7 @@ func (_m *RolePermission) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field update_by", values[i])
 			} else if value.Valid {
-				_m.UpdateBy = value.Int64
+				_m.UpdateBy = uint64(value.Int64)
 			}
 		case rolepermission.FieldDeleteTime:
 			if value, ok := values[i].(*sql.NullTime); !ok {
@@ -139,7 +139,7 @@ func (_m *RolePermission) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field delete_by", values[i])
 			} else if value.Valid {
-				_m.DeleteBy = value.Int64
+				_m.DeleteBy = uint64(value.Int64)
 			}
 		case rolepermission.FieldRoleID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
