+++ database/ent/verifycode.go
@@ -21,15 +21,15 @@ type VerifyCode struct {
 	// 创建时间
 	CreateTime time.Time `json:"create_time,omitempty"`
 	// 创建人ID
-	CreateBy int64 `json:"create_by,omitempty"`
+	CreateBy uint64 `json:"create_by,omitempty"`
 	// 更新时间
 	UpdateTime time.Time `json:"update_time,omitempty"`
 	// 更新人ID
-	UpdateBy int64 `json:"update_by,omitempty"`
+	UpdateBy uint64 `json:"update_by,omitempty"`
 	// 删除时间
 	DeleteTime time.Time `json:"delete_time,omitempty"`
 	// 删除人ID
-	DeleteBy int64 `json:"delete_by,omitempty"`
+	DeleteBy uint64 `json:"delete_by,omitempty"`
 	// 验证码
 	Code string `json:"code,omitempty"`
 	// 标识符（手机号/邮箱等）
@@ -93,7 +93,7 @@ func (_m *VerifyCode) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field create_by", values[i])
 			} else if value.Valid {
-				_m.CreateBy = value.Int64
+				_m.CreateBy = uint64(value.Int64)
 			}
 		case verifycode.FieldUpdateTime:
 			if value, ok := values[i].(*sql.NullTime); !ok {
@@ -105,7 +105,7 @@ func (_m *VerifyCode) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field update_by", values[i])
 			} else if value.Valid {
-				_m.UpdateBy = value.Int64
+				_m.UpdateBy = uint64(value.Int64)
 			}
 		case verifycode.FieldDeleteTime:
 			if value, ok := values[i].(*sql.NullTime); !ok {
@@ -117,7 +117,7 @@ func (_m *VerifyCode) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field delete_by", values[i])
 			} else if value.Valid {
-				_m.DeleteBy = value.Int64
+				_m.DeleteBy = uint64(value.Int64)
 			}
 		case verifycode.FieldCode:
 			if value, ok := values[i].(*sql.NullString); !ok {
